requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1
requests==2.32.3
anthropic==0.49.0
APScheduler==3.10.4
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.166.0
